Move LocaleGetter to locale package
